Now you can't upload transcripts with the same name ON THE BACKEND; Need to Fix FRONTEND
When you try to upload transcripts of the same name, the backend refuses to upload and sends an error with the message "A transcript with the same name already exists". Frontend should fix the webpage so that the button to visualize a transcript appears ONLY IF no error is encountered on the backend. If the backend receives an error, the frontend should not put the visualize transcript button for that transcript.